And breath... Done with cleanups. Created new app "backend", which is a huge wrapper for all the backend stuff. Flushed database clear, because it is good for further testing. Now working on merging with the master branch.
