+++ Chat/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
-from .models import QAPairEntry, TranslatorEntry
+
+from backend.models import QAPairEntry, TranslatorEntry
 
 admin.site.register(QAPairEntry)
 admin.site.register(TranslatorEntry)
