+++ backend/models.py
@@ -1,3 +1,27 @@
 from django.db import models
 
-# Create your models here.
+
+# A question and answer pair
+class QAPairEntry(models.Model):
+    question = models.IntegerField()
+    answer = models.IntegerField()
+
+    def __str__(self):
+        return "Q: " + str(self.question) + "\t" + "A: " + str(self.answer)
+
+    class Meta:
+        verbose_name = "QA Pair Entry"
+        verbose_name_plural = "QA Pairs"
+
+
+# A translator entry to translate key to value
+class TranslatorEntry(models.Model):
+    key = models.IntegerField()
+    value = models.CharField(max_length=1000)
+
+    def __str__(self):
+        return "Key: " + str(self.key) + "\t" + "Value: " + str(self.value)
+
+    class Meta:
+        verbose_name = "Translator Entry"
+        verbose_name_plural = "Translator Entries"
